+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `ProjectCard`: Add mobile styles.
 - Fix: `SummaryCard`: Fix title-bar styles.
 
 ## [6.14.0] - 2022-01-03